CrewAI/requirements.txt
langchain
langchain-core
langchain-community
crewai
langchain-groq
Pandas
crewai[tools]
python-dotenv
pyowm